Добавлено: 1. Добавление, изменение оппонентов 2. Добавление, изменение типов игр

+++ src/main/java/db/tables/game_type/DBGameTypeItem.java
@@ -59,6 +59,6 @@ public class DBGameTypeItem extends DBTableItem<DBGameType> {
         if (obj == null || obj.getClass() != this.getClass()) return false;
 
         DBGameTypeItem item = (DBGameTypeItem) obj;
-        return this.getIdGameType().getValue() == item.getIdGameType().getValue();
+        return this.getIdGameType().getValue().intValue() == item.getIdGameType().getValue().intValue();
     }
 }
